extract share
