+++ src/components/donation/index.ts
@@ -1,1 +1,2 @@
-export * from "./Steps";
+export { Steps } from "./Steps";
+export { default as Share } from "./Steps/Share";
